+++ src/test/java/com/alex/d/springbootatm/controller/ATMModelControllerTest.java
@@ -43,11 +43,16 @@ class ATMModelControllerTest {
     void testGetBalance() {
         String cardNumber = "4000007329214081";
         BankCardModel bankCardModel = new BankCardModel(1L, cardNumber, "5356", BigDecimal.valueOf(1000));
+
         when(bankCardRepository.findByCardNumber(cardNumber)).thenReturn(Optional.of(bankCardModel));
         when(atmService.checkBalance(cardNumber)).thenReturn(BigDecimal.valueOf(1000));
+
         ResponseEntity<BalanceResponse> response = atmController.getBalance(cardNumber);
+
         assertEquals(HttpStatus.OK, response.getStatusCode());
+
         assertNotNull(response.getBody());
+
         assertEquals(cardNumber, response.getBody().getCardNumber());
         assertEquals(BigDecimal.valueOf(1000), response.getBody().getBalance());
     }
@@ -57,12 +62,16 @@ class ATMModelControllerTest {
         String cardNumber = "4000007329214081";
         BigDecimal amount = BigDecimal.valueOf(500);
         BankCardModel recipientCard = new BankCardModel(1L, cardNumber, "5356", BigDecimal.valueOf(0));
+
         when(bankCardRepository.findByCardNumber(cardNumber)).thenReturn(Optional.of(recipientCard));
+
         ResponseEntity<DepositResponse> response = atmController.depositCash(cardNumber, amount);
+
         assertEquals(HttpStatus.OK, response.getStatusCode());
         assertNotNull(response.getBody());
         assertEquals(cardNumber, response.getBody().getCardNumber());
         assertEquals(BigDecimal.valueOf(500), response.getBody().getBalance());
+
         verify(bankCardRepository, times(1)).findByCardNumber(cardNumber);
         verify(atmService, times(1)).depositCashFromATM(Optional.of(recipientCard), amount);
     }
