+++ src/test/java/com/alex/d/springbootatm/service/ATMModelServiceTest.java
@@ -22,7 +22,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-@SpringBootTest
 public class ATMModelServiceTest {
 
     @Mock
@@ -40,21 +39,29 @@ public class ATMModelServiceTest {
     @Test
     public void testCheckBalance() {
         String cardNumber = "1234567890123456";
+
         BankCardModel bankCardModel = new BankCardModel();
         bankCardModel.setCardNumber(cardNumber);
         bankCardModel.setBalance(BigDecimal.valueOf(1000));
+
         when(bankCardRepository.findByCardNumber(cardNumber)).thenReturn(Optional.of((bankCardModel)));
+
         BigDecimal balance = atmService.checkBalance(cardNumber);
+
         assertEquals(BigDecimal.valueOf(1000), balance);
     }
 
     @Test
     public void testDeleteCardByNumber() {
         String cardNumber = "1234567890123456";
+
         BankCardModel bankCardModel = new BankCardModel();
         bankCardModel.setCardNumber(cardNumber);
+
         when(bankCardRepository.findByCardNumber(cardNumber)).thenReturn(Optional.of((bankCardModel)));
+
         Optional<BankCardModel> deletedCard = atmService.deleteCardByNumber(cardNumber);
+
         assertEquals((bankCardModel), deletedCard);
     }
 
@@ -62,10 +69,14 @@ public class ATMModelServiceTest {
     public void testSendTransaction() {
         BankCardModel senderCard = new BankCardModel();
         senderCard.setBalance(BigDecimal.valueOf(1000));
+
         BankCardModel recipientCard = new BankCardModel();
         recipientCard.setBalance(BigDecimal.valueOf(500));
+
         BigDecimal amount = BigDecimal.valueOf(200);
+
         atmService.sendTransaction(Optional.of((senderCard)), Optional.of((recipientCard)), amount);
+
         verify(bankCardRepository, times(1)).save(senderCard);
         verify(bankCardRepository, times(1)).save(recipientCard);
     }
@@ -73,34 +84,49 @@ public class ATMModelServiceTest {
     @Test
     void testDepositCashFromATM() {
         BankCardModel card = new BankCardModel(1L, "1234567890123456", "1111", BigDecimal.valueOf(500));
+
         BigDecimal amount = BigDecimal.valueOf(200);
+
         List<ATMModel> allATMModels = new ArrayList<>();
         allATMModels.add(new ATMModel(1L, "ATM1", "null"));
         allATMModels.add(new ATMModel(2L, "ATM2", "null"));
+
         when(atmRepository.findAll()).thenReturn(allATMModels);
         atmService.depositCashFromATM(Optional.of(card), amount);
+
         assertEquals(BigDecimal.valueOf(700), card.getBalance());
+
         verify(transactionRepository, times(1)).save(any());
     }
 
     @Test
     void testWithdrawFromATM() {
         BankCardModel card = new BankCardModel(1L, "1234567890123456", "1111", BigDecimal.valueOf(500));
+
         BigDecimal amount = BigDecimal.valueOf(200);
+
         List<ATMModel> allATMModels = new ArrayList<>();
         allATMModels.add(new ATMModel(1L, "ATM1", "null"));
+
         when(atmRepository.findAll()).thenReturn(allATMModels);
+
         atmService.withdrawFromATM(Optional.of(card), amount);
+
         assertEquals(BigDecimal.valueOf(300), card.getBalance());
+
         verify(transactionRepository, times(1)).save(any());
     }
 
     @Test
     void testCreateCard() {
         BankCardModel newCard = new BankCardModel(1L, "1234567890123456", "1111", BigDecimal.valueOf(0));
+
         when(bankCardRepository.save(any())).thenReturn(newCard);
+
         BankCardDTO createdCard = atmService.createCard();
+
         assertNotNull(createdCard);
+
         assertEquals("1234567890123456", createdCard.getCardNumber());
         assertEquals("1111", createdCard.getPinCode());
         assertEquals(BigDecimal.valueOf(0), createdCard.getBalance());
